rough merge refactor + game state (needs work)
